fix: cast average age to float in getUserAverageAge method

+++ app/Domains/User/Repositories/DashboardRepository.php
@@ -62,7 +62,7 @@ class DashboardRepository
 
     public function getUserAverageAge(): ?float
     {
-        return User::selectRaw('AVG(DATE_PART(\'year\', AGE(birth_date))) as average_age')
+        return (float) User::selectRaw('AVG(DATE_PART(\'year\', AGE(birth_date))) as average_age')
             ->value('average_age');
     }
